requirements.txt
adal==1.2.7
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
applicationinsights==0.11.10
argcomplete==3.5.3
attrs==25.3.0
azure-appconfiguration==1.7.1
azure-batch==15.0.0b2
azure-common==1.1.28
azure-core==1.33.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==1.0.0a0
azure-graphrbac==0.61.2
azure-identity==1.21.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.1.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==33.0.0
azure-mgmt-containerinstance==10.2.0b1
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==34.2.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-store==1.1.0b1
azure-mgmt-datamigration==10.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault<11.0.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-mysqlflexibleservers==1.0.0b3
azure-mgmt-netapp==10.1.0
azure-mgmt-network==28.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-postgresqlflexibleservers==1.1.0b2
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.5.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.1.0b1
azure-mgmt-servicelinker==1.2.0b3
azure-mgmt-signalr==2.0.0b2
azure-mgmt-sql==4.0.0b20
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==22.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.3.1
azure-monitor-query==1.2.0
azure-multiapi-storage==1.4.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.20.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.7.0
azureml-automl-core==1.60.0
azureml-core==1.60.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.5
azureml-dataset-runtime==1.60.0
azureml-pipeline==1.60.0
azureml-pipeline-core==1.60.0
azureml-pipeline-steps==1.60.0
azureml-sdk==1.60.0
azureml-telemetry==1.60.0
azureml-train-automl-client==1.60.0
azureml-train-core==1.60.0
azureml-train-restclients-hyperdrive==1.60.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==44.0.2
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.2.2
fabric==3.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.3.2
fusepy==3.0.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.2.0
importlib_resources==6.4.0
invoke==2.2.0
isodate==0.7.2
javaproperties==0.5.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsondiff==2.0.0
jsonpickle==4.0.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
knack==0.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
microsoft-security-utilities-secret-masker==1.0.0b4
mpmath==1.3.0
msal==1.31.2b1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pillow==11.1.0
pkginfo==1.12.1.2
portalocker==2.10.1
psutil==7.0.0
py-deviceid==0.1.1
pyarrow==19.0.1
pyasn1==0.6.1
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyGithub==1.59.1
Pygments==2.19.1
PyJWT==2.10.1
pymsalruntime==0.17.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
scp==0.13.6
SecretStorage==3.3.3
semver==2.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sshtunnel==0.1.5
starlette==0.46.2
sympy==1.13.3
tabulate==0.9.0
torch==2.2.2
torchvision==0.17.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.3.3
websockets==15.0.1
wrapt==1.17.2
xmltodict==0.14.2
zipp==3.21.0
